requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
alembic>=1.14.0
celery[redis]>=5.4.0
redis>=5.0
sse-starlette>=2.0
python-multipart>=0.0.9

# Core orchestrator deps
ollama>=0.4.0
boto3>=1.35.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
chromadb>=0.5.0
